# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  handler.MessageData:
    properties:
      message:
        example: success
        type: string
    type: object
...
  handler.UserPageData:
    properties:
      list:
        items:
          $ref: '#/definitions/service.UserDetailResponse'
        type: array
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      total:
        example: 1
        type: integer
    type: object
  response.Body:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  service.RoleItem:
    properties:
      code:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: integer
    type: object
...
  service.UserAssignRolesRequest:
    properties:
      role_ids:
        items:
          type: integer
        type: array
    type: object
  service.UserCreateRequest:
    properties:
      nickname:
        maxLength: 128
        type: string
      password:
        maxLength: 64
        minLength: 6
        type: string
      role_ids:
        items:
          type: integer
        type: array
      status:
        type: integer
      username:
        maxLength: 64
        minLength: 3
        type: string
    required:
    - nickname
    - password
    - username
    type: object
  service.UserDetailResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      nickname:
        type: string
      roles:
        items:
          $ref: '#/definitions/service.RoleItem'
        type: array
      status:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  service.UserUpdateRequest:
    properties:
      nickname:
        maxLength: 128
        type: string
      password:
        maxLength: 64
        minLength: 6
        type: string
      status:
        type: integer
    type: object
info:
  contact: {}
  description: |-
    Permission management system built with Gin, MySQL, Redis, Casbin, Cobra and slog.
    Use the login endpoint first, then put `Bearer <token>` into the Authorization header.
  title: Permission System API
  version: "1.0"
paths:
...
  /api/v1/users:
    get:
      description: List users with optional keyword and pagination filters.
      parameters:
      - description: Search keyword
        in: query
        name: keyword
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/handler.UserPageData'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Create a new user and optionally bind roles.
      parameters:
      - description: Create user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UserCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: created
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/service.UserDetailResponse'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - User
  /api/v1/users/{id}:
    delete:
      description: Delete a user by ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/handler.MessageData'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - User
    get:
      description: Get user detail by ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/service.UserDetailResponse'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Get user detail
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update user nickname, password or status.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/service.UserDetailResponse'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - User
  /api/v1/users/{id}/roles:
    put:
      consumes:
      - application/json
      description: Replace all user roles with the provided role ID list.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Assign roles request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UserAssignRolesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/service.UserDetailResponse'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Assign roles to user
      tags:
      - User
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter the JWT token with the `Bearer ` prefix, for example: Bearer
      eyJhbGciOi...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"